# Make _params_to_array faster