--- a/src/inference/Inference.jl
+++ b/src/inference/Inference.jl
@@ -242,7 +242,7 @@ end
 function _params_to_array(ts::Vector{<:AbstractTransition}, spl::Sampler)
     names_set = Set{String}()
     # Extract the parameter names and values from each transition.
-    dicts = map(enumerate(ts)) do (i, t)
+    dicts = map(ts) do t
         nms, vs = flatten_namedtuple(t.θ)
         for nm in nms
             push!(names_set, nm)
@@ -257,8 +257,8 @@ function _params_to_array(ts::Vector{<:AbstractTransition}, spl::Sampler)
     return names, vals
 end
 
-function flatten_namedtuple(nt::NamedTuple{pnames}) where {pnames}
-    temp = map(keys(nt)) do k
+function flatten_namedtuple(nt::NamedTuple)
+    names_vals = mapreduce(vcat, keys(nt)) do k
         v = nt[k]
         if length(v) == 1
             return [(string(k), v)]
@@ -268,7 +268,6 @@ function flatten_namedtuple(nt::NamedTuple{pnames}) where {pnames}
             end
         end
     end
-    names_vals = vcat(temp...)
     return [vn[1] for vn in names_vals], [vn[2] for vn in names_vals]
 end