This PR optimizes the `_params_to_array` function. In the following example, the running time of `sample` was cut down from 22 s to 8 s. Of these times, < 2 s is spent doing the sampling and the rest is saving the chain. There is a lot of room for improvement here. The 50000 iterations case terminates in 134 s now of which 16 s is sampling. Before this PR, it was too slow to even wait and time it.

```julia
using Turing

fake_dat = [5*ones(Int64, 30) for j=1:18]
fake_N = fill(5*30, 18) #Total N

@model multinomial_RE(data, N, ::Type{T}=Vector{Float64}) where {T} = begin
    B = length(data)
    R = length(data[1,1])
    theta = [T(undef, R) for j = 1:B]
    beta ~ Dirichlet(R, 1)
    alpha ~ truncated(Normal(1, 100), 0, Inf)
    for b in 1:B
    	theta[b] ~ Dirichlet(beta * alpha)
    	if any(isnan, theta[b])
    		@logpdf() = -Inf
    		return
    	end
    	data[b] ~ Multinomial(N[b], theta[b])
    end
end
model2 = sample(multinomial_RE(fake_dat, fake_N), MH(), 5000);
@time model2 = sample(multinomial_RE(fake_dat, fake_N), MH(), 5000);
```