@mohamed82008 2019-12-27
Thanks @devmotion! I have this impulse to add `@inline` next to small recursive functions I write which I should probably control :)