@mohamed82008 2020-01-05
Yes I was planning to try to make this even faster, but we can merge for now after the rebase and I will make another PR.